Fix tuple false-positive, move depth tests to own file
Track array depth per recursion chain rather than counting all '[' in the
raw type string, which previously rejected valid tuples like
(uint256[],uint256[],...) with many array-typed fields.

Move AbiTypeConverterDepthTests to its own file with file-scoped
namespace and replace LINQ usage with a StringBuilder helper.
