+++ src/Nethermind/Nethermind.Abi/AbiType.cs
@@ -84,24 +84,12 @@ namespace Nethermind.Blockchain.Contracts.Json
     {
         private const int MaxArrayDepth = 32;
 
-        private static void ValidateArrayDepth(string type)
-        {
-            int depth = 0;
-            foreach (char c in type)
-            {
-                if (c == '[' && ++depth > MaxArrayDepth)
-                    throw new AbiException(
-                        $"ABI array nesting exceeds maximum depth of {MaxArrayDepth}.");
-            }
-        }
-
         public override AbiType? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
         {
             string? type = reader.GetString()!;
-            ValidateArrayDepth(type);
-            return ParseAbiType(type);
+            return ParseAbiType(type, 0);
 
-            static AbiType ParseAbiType(string type)
+            static AbiType ParseAbiType(string type, int arrayDepth)
             {
                 // Handle plain "tuple" without array suffix
                 // Note: "tuple[]" and "tuple[N]" fall through to array handling below
@@ -112,6 +100,10 @@ namespace Nethermind.Blockchain.Contracts.Json
                 int lastBracket = type.LastIndexOf('[');
                 if (lastBracket >= 0 && type.EndsWith(']'))
                 {
+                    if (arrayDepth >= MaxArrayDepth)
+                        throw new AbiException(
+                            $"ABI array nesting exceeds maximum depth of {MaxArrayDepth}.");
+
                     string bracketContent = type[(lastBracket + 1)..^1];
                     string elementTypeStr = type[..lastBracket];
 
@@ -120,13 +112,13 @@ namespace Nethermind.Blockchain.Contracts.Json
                         case > 0 when int.TryParse(bracketContent, out int length):
                             {
                                 // Fixed-size array: type[N]
-                                AbiType elementType = ParseAbiType(elementTypeStr);
+                                AbiType elementType = ParseAbiType(elementTypeStr, arrayDepth + 1);
                                 return new AbiFixedLengthArray(elementType, length);
                             }
                         case 0:
                             {
                                 // Dynamic array: type[]
-                                AbiType elementType = ParseAbiType(elementTypeStr);
+                                AbiType elementType = ParseAbiType(elementTypeStr, arrayDepth + 1);
                                 return new AbiArray(elementType);
                             }
                         default:
@@ -153,7 +145,7 @@ namespace Nethermind.Blockchain.Contracts.Json
                 AbiType[] abiTypes = new AbiType[types.Length];
                 for (int i = 0; i < types.Length; i++)
                 {
-                    abiTypes[i] = ParseAbiType(types[i].Trim());
+                    abiTypes[i] = ParseAbiType(types[i].Trim(), 0);
                 }
 
                 return new AbiTuple(abiTypes);
